cpuidle: t18x: Fix up driver for k4.4
The CPUIDLE_FLAG_TIME_VALID was removed on k4.3 and future kernels and
every driver is assumed to have that be the case by default now.

Also fixed up API changes for __{enable/disable}_irq that drop the irq
parameter because it can be fetched from the irq descriptor instead of
being passed as an arg.

Both these changes are purely cosmetic, no functional change on either
of k3.18 and k4.4.

Change-Id: Iefbe7dfb61c83f9dfb19e91c7a20d0ba3bb28dbe
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: http://git-master/r/1140625
Reviewed-by: Krishna Sitaraman <ksitaraman@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>