video: tegra: nvmap: fail ivc duplication when handle's ref count is zero
handle's ref count shouldn't be incremented when it is zero.
So, such a handle can't be duplicated. Fail the duplication.

However, handle's heap block is not yet freed. Strip this block
so that the create_from_ivc ioctl can use this buffer to create
a new handle from it.

If the handle's heap block is unused and left to get free, then
create_from_ivc could fail when create is attempted before heap
block's free. This is not expected from create_from_ivc.

bug 1757413

Change-Id: I5efd2a5f0147ba5875735f0afcb0ca3f289ec517
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/1133975
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Tested-by: Krishna Reddy <vdumpa@nvidia.com>