spi: tegra: Add SPI proxy driver for AON SPI
SPI2 is accessible by both AON/SPE and CCPLEX. If both CPU's access the SPI
bus at the same time, there will be a conflict. In order to ensure there are
no conflicts, we select SPE as the owner of the SPI2 controller and all the
other CPU's that need to touch the SPI2 controller will send an IPC message
to the SPE and SPE will perform the SPI transaction on the requesting CPU's
behalf. This driver registers as the standard kernel SPI controller driver
for AON SPI controller i.e. SPI2. This driver sends the IPC messages comprising
of the SPI transfers to SPE instead of directly programming the SPI registers.

Bug 1701661

Change-Id: I149d2bdd938c3f3aa03a754c2cc0678758fbba9e
Signed-off-by: Akhilesh Reddy Khumbum <akhumbum@nvidia.com>
Reviewed-on: http://git-master/r/1117289
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Krishna Yarlagadda <kyarlagadda@nvidia.com>
Tested-by: Krishna Yarlagadda <kyarlagadda@nvidia.com>
Reviewed-by: Mustafa Bilgen <mbilgen@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>