arm64: dts: Update cpu-capacities for A57s on T18x
The relative CPU capacity of the A57s compared to D15s is now different
because the Fmax's of the two clusters changed. With the current Fmax of
2.1GHz on A57s and 2.2GHz on D15s, with a 1.3 IPC advantage for D15s, we
get a relative capacity value of 752 for A57s.

Reflecting the true capacities of the CPUs will help when scaling
scaling frequencies as load moves across clusters. It also potentially
boosts multi-threaded performance in overloaded scenarios by accurately
reflecting the CPU capacities of the A57s.

Change-Id: I56bde144388491867fb70dbbffe22c929bd80bd9
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: http://git-master/r/1142968
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Peter Boonstoppel <pboonstoppel@nvidia.com>
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>