tegra: add HSP (mailbox) platform driver
This adds a "proper" platform device for Tegra 186 HSP blocks.

Currently, only shared mailboxes are handled. The existing HSP
doorbell support is left in the existing pseudo-driver for
compatibility with early boot and BPMP. Fortunately, the two
drivers handle distinct registers and interrupts (except for
both reading the dimensioning register).

Mailboxes are assumed to be always used in (even, odd) pair. This is as
per the Tegra hardware synchronization IAS. Indeed, a pair of mailboxes
sits on a single aligned 64 KiB page.

CRTC-143

Signed-off-by: Remi Denis-Courmont <remid@nvidia.com>
Change-Id: I88bec1dcc8b49b19bdd08185e6c517acd0c117d2
Reviewed-on: http://git-master/r/1143209
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mika Liljeberg <mliljeberg@nvidia.com>