hsp: do not assume single HSP instance
There are multiple Tegra HSP instances on the SoC. In Tegra HSP
doorbell driver needs to match the "TOP0" instance, which contains
the doorbells. So far, it worked because other instances were not
defined in the Device Tree.

This patch makes the doorbell driver look for the HSP instance with
a doorbell IRQ. This also fixes a leak on error and gets rid of
unnecessary properties from Device Tree.

Signed-off-by: Remi Denis-Courmont <remid@nvidia.com>
Change-Id: I0385dfe562f2f4e861b0f7460aa3be2246dc17e4
Reviewed-on: http://git-master/r/1143277
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mika Liljeberg <mliljeberg@nvidia.com>