tegra: dc: add missing SD ifdef checks
This panel file assumes that the SD2.0 driver is compiled.
That is no longer the case after being moved to kernel-display,
since this panel file is also being compiled for t186 where
SD2.0 doesn't exist. Added checks for NVDISPLAY and calls
to SD2.0 or NvDisp SD drivers as appropriate.

Bug 1761462

Change-Id: I63a9d131ebc4f1ef43f68bd72c3ea76e61422fac
Signed-off-by: Daniel Solomon <daniels@nvidia.com>