vi: fix fallback to host-based VI Notify
We cannot use both VI Notify drivers simultaneously. So far, the driver
selection was done by looking for the rtcpu-vinotify top-level device
node in device tree. But on the one hand, the implementation of the
check was flaky, and on the other hand, rtcpu-vinotify node is removed.

This uses the Camera RTCPU device alias instead, and also fixes
surrounding error handling.

JIRA CRTC-149

Signed-off-by: Remi Denis-Courmont <remid@nvidia.com>
Change-Id: I3a17cf30b3782458466865e9e3f5da29c469d20f
Reviewed-on: http://git-master/r/1149089
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mika Liljeberg <mliljeberg@nvidia.com>