ivc: fix missing peer notification
Ensure that we notify the remote end each time our local
state has changed. Previously, we neglected notificaton if
the remote end was already in the established state, but
this optimization can lead to a hang in the protocol if
the remote end is undergoing a state change.

VFND-1786

Change-Id: I1de31d8453ec7ba32d3b79bf4bd40f7edf41aeed
Reviewed-on: http://git-master/r/1149233
Reviewed-on: http://git-master/r/1198265
Reviewed-on: http://git-master/r/1323006
GVS: Gerrit_Virtual_Submit
Reviewed-by: Aniruddha Banerjee <aniruddhab@nvidia.com>
Tested-by: Aniruddha Banerjee <aniruddhab@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>