video: tegra: host: acquire mutex always in abort path
In cdma_handle_timeout(), we currently have mutex_trylock()
to acquire the cdma lock.

In case of real timeout, if lock is already held in other
thread (e.g. work completion), we detect it using
mutex_trylock() and then re-schedule the timeout

But this is not necessary in channel_abort() path
In channel_abort() patch, we need to ensure that we clean
up the channel in same thread without rescheduling

Hence use mutex_lock() if skip_reset is set (abort path)
or otherwise continue using mutex_trylock()

Bug 1695878

Change-Id: Ib434d2cf3b04f179dbd5196331f6bc816778dff7
Signed-off-by: Deepak Nibade <dnibade@nvidia.com>
Reviewed-on: http://git-master/r/1151821
GVS: Gerrit_Virtual_Submit
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>