platform: tegra: pm: misc fixes for wake logic
Reading and writing wake levels was not performed correctly.
Reading due to the reg increment being in the wrong loop,
causing the wrong register to be read.
Writing due to that when modifying the wake level register,
bits were only set, never cleared.

Clearing the status registers was not performed correctly.
The HW status register clear function didn't use the correct
register to mask which bits to clear.
The SW status register clear was issuing writes to the wrong
register.

Finally an off by one bug is fixed in the print_vals() function.

Bug 1737058

Change-Id: I1c1e6a4c79501df3f3ad051f79c537e97ef45c0f
Signed-off-by: Prashant Gaikwad <pgaikwad@nvidia.com>
Reviewed-on: http://git-master/r/1152437
Reviewed-by: Stefan Kristiansson <stefank@nvidia.com>
Tested-by: Stefan Kristiansson <stefank@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Timo Alho <talho@nvidia.com>