rtcpu: support intermediate FW
The intermediate firmware does not require kernel to do
synchronization. Synchronization is done only if there is a dummy
message in smbox 7 when RTCPU is booted.

Once kernel is updated, we update firmware to a version requiring the
synchronization. After the final firmware is submitted, this change can
be reverted.

JIRA CRTC-135
Bug 1699448
Bug 200177745

Change-Id: Ia4ae6f55d1ff29d91d3d221e9a0ce75e5543de5f
Signed-off-by: Pekka Pessi <ppessi@nvidia.com>
Reviewed-on: http://git-master/r/1153610
Reviewed-by: Remi Denis-Courmont <remid@nvidia.com>
Reviewed-by: Mika Liljeberg <mliljeberg@nvidia.com>