rtcpu/debug: remove parameters mutex
The parameters mutex was useless. If userspace modifies a parameter while
it is being used, then userspace is broken. The kernel does not really
care: it will use some value, and this does not affect security.

The transaction mutex is made interruptible, so a process
waiting for its turn can be interrupted or killed.

Also the completion timeout is made killable, so that a process stuck
in a failed transaction can at least be killed. Otherwise, it would
potentially remain in D state for very long.

Finally, fix error handling in debugfs initialization and use built-in
debugfs helpers for parameters.

JIRA CRTC-253

Signed-off-by: Remi Denis-Courmont <remid@nvidia.com>
Change-Id: I1c0fdaeb8c7dd38a40ef3c66a9b295c079e29eb0
Reviewed-on: http://git-master/r/1153461
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Pekka Pessi <ppessi@nvidia.com>
Reviewed-by: Mika Liljeberg <mliljeberg@nvidia.com>