rtcpu: remove racy notify at IVC channel probe
At this point, none of the IVC channel protocols involve RTCPU sending
unsolicited messages. CCPLEX is always the first side to send. So the
IVC channel buffers are always empty in both directions when the
channel driver gets probed, and this notification is useless.

Going forward, if that assumption no longer holds, the IVC channel
driver will have to manually check (with tegra_ivc_can_read()) the
initial buffer status. The generic IVC bus is not in a position to
perform the necessary synchronization/locking at things currently
stand. Namely, nothing prevents an IRQ from HSP and a channel probing
from occurring concurrently, leading to races as tegra_hsp_notify()
is not reentrant.

Jira CRTC-253

Signed-off-by: Remi Denis-Courmont <remid@nvidia.com>
Change-Id: I8c286b48ccc4b5a1a0d7ac4512f031e64e15005c
Reviewed-on: http://git-master/r/1159315
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mika Liljeberg <mliljeberg@nvidia.com>