This driver will basically translate serial communication to i2c communication between the user-space and the GPS module.
It creates a /dev/ttyS device node.

There are specific tty termios flags in order to the tty line discipliner not
to change the date it is pushed to user space. That is necessary so the NMEA
data is not corrupted.
* iflags: added IGNCR: Ignore carriage return on input.
* oflags: removed ONLCR: (XSI) Map NL to CR-NL on output.

Change-Id: Ib03c544a408f73442f0e36c3626e05a82a33d580
Signed-off-by: Felipe F. Tonello <eu@felipetonello.com>

Bug 1779597

Signed-off-by: Rakesh Goyal <rgoyal@nvidia.com>
Change-Id: Ib03c544a408f73442f0e36c3626e05a82a33d580
Reviewed-on: http://git-master/r/1195351
(cherry picked from commit 723bb79b25049590bb091e728a4a84216be7c143)