hid: jarvis: On timer_start, do not delete active timer
Whenever the try_to_del_timer_sync call returns with error,
the decoding_timer that is supposed to decode the audio
packets does not get scheduled and hence the FIFO gets full
(no consumer to consume the packets in the fifo). The cases
in which this happens is when the timer is active when
try_to_del_timer_sync is called. We do not need to remove
the timer when starting voice search. snd_atvr_schedule_timer
calls mod_timer(), which just modifies the expiration time
of the timer if its already active and schedules one if its
not already active.

Bug 200195718

Change-Id: Iec728dfd7eb002251afa8a7923ab2e2167d718d9
Signed-off-by: Ankita Garg <ankitag@nvidia.com>
Reviewed-on: http://git-master/r/1160526
(cherry picked from commit 1909d374bf303051f3949f68d58c7bf509f76c87)
Reviewed-on: http://git-master/r/1162699
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>