cpufreq: t18x: Fix potential memory corruption
When allocating the frequency tables, we try to allocate

	(ndiv_max - ndiv_min) / freq_table_step_size

number of elements. However, we store both ndiv_min AND ndiv_max in said
table, so we want one more element to cover for that..

This fixes a potential memory corruption in the kernel given that we end
up exceeding the number of allocated elements with index which would
potentially then clobber someone else's memory.

Change-Id: I3c43d19c8134f9759449b1e0169fd32d2fcd1330
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: http://git-master/r/1162093
GVS: Gerrit_Virtual_Submit
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>