platform: t18x: add stacked IRQ domain support to pm-irq
In k4.4 it is no longer allowed to (ab)use gic_arch_extn to
provide wake extensions to the GIC, instead a separate
irqchip has to be created which serves as a child of the
GIC.
This child get assigned the IRQs that has wake functionality
and implements the appropriate set_wake ops for that.
All other ops for these IRQs are passed to the parent (GIC).

Backwards compatibility with k3.18 has been mantained and is
guarded behind the CONFIG_IRQ_DOMAIN_HIERARCHY knob (defined
for k4.4, not defined for k3.18).

One small caveat is that the hardware IRQ numbers are
mapped with an offset of 32, this is to match the behaviour
of the GIC mappings.
By doing this, it is not necessary
to make changes to the wake<->irq mappings in the wakeup
driver, and we can simply use the hwirq to obtain the wake
number from tegra_irq_to_wake() for both k3.18 and k4.4.
(In k4.4 it is not possible to use the "virtual" irq,
since the assumption of a 1:1 mapping between hwirq and
"virtual" irq does not necessarily hold).

Bug 200193765

Change-Id: I09dd0152e85f9cb8a8bed1ce9c4581202d18d13b
Signed-off-by: Stefan Kristiansson <stefank@nvidia.com>
Reviewed-on: http://git-master/r/1164769
GVS: Gerrit_Virtual_Submit
Reviewed-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>