video: tegra: dc: fix timestamp handling
This change resolves an issue where timestamp
node was not getting added to the list for the
first frame. It also fixes an issue where
timestamp node is deleted from the list when
the node itself may not be in the list to begin
with, leading to kernel panic.

Bug 1775101

Change-Id: I2aed6c868ac135688091d953083a925fd2445d4f
Reviewed-on: http://git-master/r/1164525
(cherry picked from commit 6df1cf5eb771afc861eaa9e65a91d5c5146d30c9)
Signed-off-by: Sachit Kadle <skadle@nvidia.com>
Reviewed-on: http://git-master/r/1166061
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
Tested-by: Jon Mayo <jmayo@nvidia.com>