video: tegradc: nvdisp: fix input_lut
When TEGRA_DC_EXT_SET_LUT ioctl is called for a window
which is disabled, programming the lut_base_addr has no
effect as the window is disabled.

Following that when we enable a window, we enable input lut
but dont program the lut_base_addr. This causes iommu errors.

This patch fixes this issue by programming the lut_base_addr
before enabling input lut when window is enabled.

Bug 1741658
TDS-190

Change-Id: Iac6d078415c913aa8dfc53eb882a2b3653886dd2
Signed-off-by: Rohit Khanna <rokhanna@nvidia.com>
Reviewed-on: http://git-master/r/1165960
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: Aron Wong <awong@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>