tegra: rtcpu: use phandle for IVC channels layout
To avoid duplicating an increasingly large list of IVC channels,
move it out of the RTCPU device node. This has no compatible property
since we do not want a platform device for it.

At the same time, this stores the AST region index in the Device Tree
instead of hard-coding it within the driver.

Device Tree bindings documentation is updated accordingly.

Jira CRTC-300

Signed-off-by: Remi Denis-Courmont <remid@nvidia.com>
Change-Id: I16dbe7af124e96f0eab0cc477994e196d73fc05e
Reviewed-on: http://git-master/r/1167011
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Pekka Pessi <ppessi@nvidia.com>
Reviewed-by: Mika Liljeberg <mliljeberg@nvidia.com>