gpu: nvgpu: Handle allocators with a base of 0
When an allocator is created with a base of 0 the first allocated
block could well be 0. This appears to be an error since gk20a_balloc()
normally returns 0 for error cases.

This patch removes one block from the allocatable resources when base
is set to 0 so that code using gk20a_balloc() does not get tricked
into thinking valid allocations are OOM cases.

Change-Id: I641642d3f790c4c7860d0d1381f4db6f4f72e709
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/1169764
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Tested-by: Terje Bergstrom <tbergstrom@nvidia.com>