video: tegra: host: always do mlock recovery in virtualization
The timeout procedure cleans everything
except MLOCK ownerhip and sends a dummy job.
It tries acquiring MLOCK so MLOCK timeout
interrupt triggers in vmserver. Vmserver moves
the ownership to the new channel who just
releases it. The result it that everything
is cleaned and MLOCK is free.

VFND-1980

Change-Id: I4fb9b01343f9d86d364a19d4c6c09d99292b9dac
Signed-off-by: Damian Halas <dhalas@nvidia.com>
Reviewed-on: http://git-master/r/1172087
GVS: Gerrit_Virtual_Submit
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>