tegra: watchdog: remove second watchdog
Linux only needs / uses one watchdog. Linux has its own per core watchdog
that detects when an individual core is no longer responing to interrupts.
The hardware watchdog is then only useful for when none of the cores are
responding.

To Linux, T186 has 6 cores. The fact that there are two clusters is not
important to how it handles interrupts.

So, instead of having a watchdog timer per cluster, only use one.

Bug 1782487
Bug 1776680

Change-Id: I569b2af6357dc100ac64d980bcbe3250ec7884df
Signed-off-by: Alex Van Brunt <avanbrunt@nvidia.com>
Reviewed-on: http://git-master/r/1172099
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Krishna Yarlagadda <kyarlagadda@nvidia.com>
Tested-by: Krishna Yarlagadda <kyarlagadda@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1209726
Reviewed-by: Venkat Reddy Talla <vreddytalla@nvidia.com>
Tested-by: Venkat Reddy Talla <vreddytalla@nvidia.com>