WAR: video: tegra: nvdisp: override hubclk rate
When BL enables display, they set the minimum required hubclk rate
for their use case. IMP will ultimately be responsible for setting
the required hubclk rate in the kernel, but it's not ready for client
adoption yet. As such, we currently don't adjust the hubclk rate,
which means that we'll be left with whatever rate BL sets in the cases
where BL display is enabled. A simple workaround is to explicitly set
the hubclk rate to 408MHz in the kernel.

Bug 200212319

Change-Id: I93f3202178b1795c5305f95e2cb197ae0f7c70d4
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1172678
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Karthik Gunda <kgunda@nvidia.com>
Reviewed-by: Manoj Gupta <manogupta@nvidia.com>
Tested-by: Manoj Gupta <manogupta@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>