video: tegra: nvdisp: cache current dc mode
When system resumes from sleep, we wait for userspace to unblank display
with a mode chosen by userspace. When there's no userspace entity, DC
stays disabled as no unblank call is generated after resume. To overcome
this, cache the previously set DC mode so that we can enable DC with the
cached mode during resume.

bug 200205112

Change-Id: I9d89e6242d15c29ef79140feb0647cdad9eef451
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: http://git-master/r/1173053
Reviewed-by: Bibek Basu <bbasu@nvidia.com>