video: tegra: dp: Fix of EDID checksum error
EDID checksum error was not passing
certification test as it was not falling
back to 640 X 480 60Hz.
Issue was in refresh rate check for 60Hz.
Because of conversion we were losing accuracy in
refresh rate and resulting in 59Hz hence the
check was getting failed and it was not doing LT
or falling back to 640 X 480.

Bug 200205365

Change-Id: I368b64668bcd52978515a4786c59c09c74d0b344
Signed-off-by: Pradeepan Patra <ppatra@nvidia.com>
Reviewed-on: http://git-master/r/1177667
Reviewed-by: Venu Byravarasu <vbyravarasu@nvidia.com>
Tested-by: Venu Byravarasu <vbyravarasu@nvidia.com>