trusty: disable Trusty's FIQ handler
This patch disables Trusy's FIQ handler for ARM64 CPUs. The
handler expects Trusty to handle the FIQ and in turn inform
the linux driver. Platforms might not want to enable Trusty but
still enable the FIQ handling functionality. A common FIQ handler
and glue would be added instead which is not dependent on the
presence of a Trusted OS on the platform.

Bug 200216015

Change-Id: I0cf9595b6b0a14e1aabfea32f9ab2cb8598bcbf0
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Reviewed-on: http://git-master/r/1178669
(cherry picked from commit 0646aafc60a1921145b1fbedb2baf178ea4420d3)