drivers: powergate: use new powergate ABI
This switches the powergate driver from the old MRQ_PG_UPDATE_STATE
ABI to the new MRQ_PG ABI.

The new MRQ_PG ABI does not do reference counting in BPMP-FW,
it only keep track of the on/off state of the domain.
So, in addition to the switch to the new ABI, reference counting
is also being added.

Bug 1833048

Change-Id: I15f009bca07f3824308ee9275a6e2aa22c67924c
Signed-off-by: Timo Alho <talho@nvidia.com>
Reviewed-on: http://git-master/r/1260008
Reviewed-by: Stefan Kristiansson <stefank@nvidia.com>
Tested-by: Stefan Kristiansson <stefank@nvidia.com>
GVS: Gerrit_Virtual_Submit