t18x: dc: add color expansion for certain color formats
Since the DC pipeline in t186 is 16 bit, we need to change
the logic for enabling color expansion. In fact, there needs
to be restriction since color expansion is a no-op for 16
bits and above.

Currently we are enabling color expansion for the following modes:
RGB full->limited range
YUV 444/422

Changing this for all RGB formats will be done in a follow on patch after changes to golden CRCs in tests

Bug 200206336

Change-Id: If1753563e87ce26edb580d823f6bc2252a8d2576
Signed-off-by: Anshuman Nath Kar <anshumank@nvidia.com>
Reviewed-on: http://git-master/r/1185268
Reviewed-by: Mitch Luban <mluban@nvidia.com>