firmware: tegra: remove ping from bpmp_connect()
When the __bpmp_do_ping() call happens, mail_inited is still 0, but the
API passed because tegra_bpmp_send_receive_atomic() does not check this
flag now (which is a bug). So, remove the call from here. We are moving
the initial ping to the driver's _probe().

Bug 1677466

Change-Id: I64079af71a5e524c9026efdf9ea6e69acd851492
Signed-off-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-on: http://git-master/r/1194571
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Stefan Kristiansson <stefank@nvidia.com>