padctl: increase delay for usb2 pad cfg capture
USB2 pad configuration programming includes two steps:
1. Capture step: capture the pad input values and saves them.
2. Drive step: start driving the saved values.

The capture step involves running the capture clock for one cycle
and saving the values in flops. Once the CAP_CFG is programmed, the register
write is moved from the APB clock domain to the slow 32KHz clock domain,
and then that is used to run the clock for one cycle. So, overall this
can take up to 5-6 cycles of 32KHz clock, that is why we need ~150-200us
delay between the two.

Bug 200199085

Change-Id: I3172699dc656a2aa0fb51bf30920370eed281fcc
Signed-off-by: Henry Lin <henryl@nvidia.com>
Reviewed-on: http://git-master/r/1201200
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>