t18x: cpufreq: Prevent hotplug and cpufreq deadlock
There currently is an ABBA deadlock with the cpu_hotplug.lock and the
policy->rwsem as follows:

cpufreq_update_policy                        -> A down_write(policy->rwsem)
    tegra_get_speed
        get_online_cpus
            mutex_lock(cpu_hotplug.lock)     -> B

_cpu_up
    cpu_hotplug_begin                        -> B mutex_lock(&cpu_hotplug.lock)
        cpufreq_cpu_callback
            cpufreq_online
                down_write(policy->rwsem)    -> A


The deadlock was specifically introduced due to tegra_get_speed() trying
to prevent a CPU from going offline before it did the speed measurement on
that CPU.

However, there is nothing protecting a CPU from going offline BEFORE the
get_online_cpus() in tegra_get_speed() so callers of driver->get() already
should be taking care of hotplug serialization if needed.

Furthermore, the smp_call_function_single() correctly checks whether a CPU
is online or not so we don't need to add this check in tegra_get_speed()
therefore we can safely drop the {get,put}_online_cpus in tegra_get_speed
and fix this deadlock.

Change-Id: I0397ccb24387586f8e29efbdf69c15d3d5b353ca
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: http://git-master/r/1201380
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>