drivers: iio: Add NVS-LED-TEST driver source
This is a driver which creates a test sensor used for latency
measurements. It simulates a sensor triggering an event through the
nvs framework. The event is triggered by a write to its sysfs "on_ms"
file. When the event is triggered, it also toggles the GPIO specified
in its DT entry to turn on an LED. This can be used with a high speed
camera to capture the event start time when doing latency measurements.

Based on work by: Andrew chew <achew@nvidia.com>

Bug 200237272
AV-133

Change-Id: Ib4f0a831b2a3da1321d48112fc66c909809105f7
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-on: http://git-master/r/1271136
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>